Refactor views and add new Welcome and StreamMap views
Renamed `WordCloudView` to `StreamCloudView` and `ThoughtStreamView` to `ThoughtView`. Introduced `WelcomeView` for onboarding and `StreamMapView` for map-based visualizations. Updated project file references accordingly.
